refactor(api): use association name for rna siren identifier
